Refactor main.rb to remove cyclometic complexity, and create app.rb
